+++ tomcat/src/main/java/nextstep/jwp/controller/LoginController.java
@@ -8,7 +8,7 @@ import nextstep.jwp.db.InMemoryUserRepository;
 import nextstep.jwp.model.User;
 import nextstep.jwp.util.QueryStringUtil;
 import org.apache.coyote.http11.cookie.Cookie;
-import org.apache.coyote.http11.handler.Controller;
+import org.apache.catalina.controller.Controller;
 import org.apache.coyote.http11.request.HttpMethod;
 import org.apache.coyote.http11.request.HttpRequest;
 import org.apache.coyote.http11.response.HttpResponse;
