+++ tomcat/src/main/java/nextstep/jwp/controller/LoginPageController.java
@@ -6,7 +6,7 @@ import static org.apache.coyote.http11.response.HttpResponseHeader.LOCATION;
 
 import java.io.IOException;
 import org.apache.coyote.http11.cookie.Cookie;
-import org.apache.coyote.http11.handler.Controller;
+import org.apache.catalina.controller.Controller;
 import org.apache.coyote.http11.request.HttpMethod;
 import org.apache.coyote.http11.request.HttpRequest;
 import org.apache.coyote.http11.response.ContentType;
