+++ tomcat/src/main/java/org/apache/coyote/http11/Http11Processor.java
@@ -6,8 +6,8 @@ import java.io.InputStreamReader;
 import java.net.Socket;
 import nextstep.jwp.exception.UncheckedServletException;
 import org.apache.coyote.Processor;
-import org.apache.coyote.http11.handler.Controller;
-import org.apache.coyote.http11.handler.ControllerMapper;
+import org.apache.catalina.controller.Controller;
+import org.apache.catalina.controller.ControllerMapper;
 import org.apache.coyote.http11.handler.ResourceHandler;
 import org.apache.coyote.http11.handler.ResourceHandlerMapper;
 import org.apache.coyote.http11.request.HttpRequest;
